e54 working with the screen for the first time ever! it is still buggy but huge milestone

+++ software/firmware/project_oracle_test_firmware/e54/e54_gfx_learning/e54_gfx_learning/drivers/p_gpio.h
@@ -17,9 +17,9 @@ typedef struct p_port_config
 
 void p_gpio_init(void);
 
-void p_gpio_parallel_write(PortGroup* group, uint32_t mask, uint32_t data);
+void p_gpio_parallel_write(PortGroup* group, uint32_t mask, uint16_t data);
 
-void p_gpio_parallel_write_arr(PortGroup* group, uint32_t mask, uint32_t* data, uint32_t len);
+void p_gpio_parallel_write_arr(PortGroup* group, uint32_t mask, uint16_t* data, uint16_t len);
 
 void p_gpio_set_port_data(PortGroup* const port, const uint32_t mask, const uint32_t data);
